editors - polish readonly handling
