+++ src/vs/workbench/browser/parts/editor/baseEditor.ts
@@ -78,11 +78,9 @@ export abstract class BaseEditor extends Panel implements IEditor {
 	 * The provided cancellation token should be used to test if the operation
 	 * was cancelled.
 	 */
-	setInput(input: EditorInput, options: EditorOptions | undefined, token: CancellationToken): Promise<void> {
+	async setInput(input: EditorInput, options: EditorOptions | undefined, token: CancellationToken): Promise<void> {
 		this._input = input;
 		this._options = options;
-
-		return Promise.resolve();
 	}
 
 	/**
